Move HTTP stream plumbing into the WordPress package
