+++ src/WordPress/DataSource/PlaygroundFetchSource.php
@@ -7,8 +7,8 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\HttpClient\Response\StreamWrapper;
 use Symfony\Contracts\EventDispatcher\Event;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
-use WordPress\Streams\StreamPeeker;
-use WordPress\Streams\StreamPeekerContext;
+use WordPress\Streams\StreamPeekerWrapper;
+use WordPress\Streams\StreamPeekerData;
 
 class PlaygroundFetchSource extends BaseDataSource {
