+++ src/WordPress/Streams/StreamPeekerData.php
@@ -2,8 +2,9 @@
 
 namespace WordPress\Streams;
 
-class StreamPeekerContext {
+class StreamPeekerData extends VanillaStreamWrapperData {
 
 	public function __construct( public $fp, public $onChunk = null, public $onClose = null ) {
+		parent::__construct( $fp );
 	}
 }
