+++ src/WordPress/Streams/VanillaStreamWrapperData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace WordPress\Streams;
+
+class VanillaStreamWrapperData {
+	public $fp;
+
+	public function __construct( $fp ) {
+		$this->fp = $fp;
+	}
+}
